ai-vocabulary-learning-platform/backend/requirements.txt
flask
psycopg2-binary
Werkzeug
flask-cors
PyJWT
together
requests
